Export a single function.
The forward qualification is not necessary. If you want a reverse
iterator, reverse the array you pass to the constructor.

I do not feel that it is necessary to maintain consistency with the
other iterator modules that have a both a `forward` and `reverse`
function, but I was thinking along those lines when I created this
module.

Closes #15.