Add LGPL 3.0.
Advance is part of a collection of MVCC utilities used to implement
durable databases on top of the Strata b-tree. Strata will be released
under the LGPL 3.0. All the supporting libraries will also be released
under LGPL 3.0. At the time of this commit, no one is using Advance and
I am the sole copyright holder, so this license applies not only to this
version but to all prior versions of Advance, regardless of the license
files that are captured by the version control software used to manage
the editing of this project.

Closes #18.