Make reversal an argument.
Make reversal an argument instead of having a forward and reverse
function. You're fussing over a single conditional at construction. With
Riffle it's even sillier because you usually build those with an
argument that determines forward or reverse, so you have to have a
branch of some sort to chose which function.

Closes #80.